github_action/requirements.txt
bs4
lxml
boto3
psycopg2-binary
python-dotenv
pillow
requests
discord_webhook
jinja2
